requirements.txt
flask
flask-socketio
pynostr
websockets
ecdsa
bech32
python-dotenv
numpy
opencv-python
rich
cryptography
